BCD/requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
attrs==23.2.0
bcrypt==4.1.2
boto3==1.34.34
botocore==1.34.34
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.3
coloredlogs==15.0.1
contourpy==1.2.0
cryptography==42.0.0
cycler==0.12.1
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gast==0.5.4
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.1
h11==0.14.0
h5py==3.10.0
humanfriendly==10.0
idna==2.8
imageio==2.33.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
largestinteriorrectangle==0.2.0
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.41.1
Mako==1.3.1
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.17.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.0
protobuf==4.25.2
psutil==5.9.8
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyMatting==1.1.12
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rembg==2.0.54
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
s3transfer==0.10.0
scikit-image==0.22.0
scikit-learn==1.4.0
scipy==1.12.0
seaborn==0.13.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
stitching==0.5.3
sympy==1.12
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2024.1.30
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.1.14
urllib3==2.0.7
uvicorn==0.27.0
watchdog==3.0.0
Werkzeug==3.0.1
wrapt==1.14.1
